Add TestRunner which supports most Centroid and ConvexHull cases.
Centroid currently fails some test against master, but I will address
this in an upcoming PR to geo.

Plus we skip a couple of tests because we don't yet have a way to handle
collections of "POINT EMPTY".

The ConvexHull cases currently fail, at least some of which is because
geo always returns a Polygon for ConvexHull, whereas JTS might return
other geometry types